Refactor, pull apart code
